[Refactor] 상태관리 파일 구조 재정의
